Fixed the internal txn ids for the protodtxncoordinator to be 64-bit. I think we're in business!!!
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/protorpc-messaging-branch@1641 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/dtxn/protodtxn/protodtxncoordinator.h
@@ -33,13 +33,13 @@ public:
             ::google::protobuf::Closure* done);
 
     // INTERNAL: Remove request from the transaction map and delete state.
-    ProtoDtxnCoordinatorRequest* internalDeleteTransaction(int32_t transaction_id);
+    ProtoDtxnCoordinatorRequest* internalDeleteTransaction(int64_t transaction_id);
 
 private:
     dtxn::DtxnManager* dtxn_manager_;
     int num_partitions_;
 
-    typedef base::unordered_map<int32_t, ProtoDtxnCoordinatorRequest*> TransactionMap;
+    typedef base::unordered_map<int64_t, ProtoDtxnCoordinatorRequest*> TransactionMap;
     TransactionMap transactions_;
 };
